Start reducing number of compilation units (trivial cases, first)
